littleeye_project/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
arrow==1.2.3
    # via -r requirements.in
asgiref==3.6.0
    # via django
babel==2.12.1
    # via mkdocs-git-revision-date-localized-plugin
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   mkdocs
    #   proselint
    #   uvicorn
colorama==0.4.6
    # via mkdocs-material
crispy-bootstrap5==0.7
    # via -r requirements.in
csscompressor==0.9.5
    # via mkdocs-minify-plugin
django==4.1.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.9.0
    # via -r requirements.in
future==0.18.3
    # via proselint
ghp-import==2.1.0
    # via mkdocs
gitdb==4.0.10
    # via
    #   gitdb2
    #   gitpython
gitdb2==4.0.2
    # via -r requirements.in
gitpython==3.1.31
    # via mkdocs-git-revision-date-localized-plugin
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
htmlmin2==0.1.13
    # via mkdocs-minify-plugin
idna==3.4
    # via requests
jinja2==3.1.2
    # via
    #   mkdocs
    #   mkdocs-material
jsmin==3.0.1
    # via mkdocs-minify-plugin
markdown==3.3.7
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markupsafe==2.1.2
    # via jinja2
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.4.2
    # via
    #   -r requirements.in
    #   mkdocs-autolinks-plugin
    #   mkdocs-exclude-search
    #   mkdocs-git-revision-date-localized-plugin
    #   mkdocs-material
    #   mkdocs-minify-plugin
    #   mkdocs-section-index
mkdocs-autolinks-plugin==0.7.0
    # via -r requirements.in
mkdocs-exclude-search==0.6.5
    # via -r requirements.in
mkdocs-git-revision-date-localized-plugin==1.2.0
    # via -r requirements.in
mkdocs-material==9.1.3
    # via -r requirements.in
mkdocs-material-extensions==1.1.1
    # via mkdocs-material
mkdocs-minify-plugin==0.6.4
    # via -r requirements.in
mkdocs-section-index==0.3.5
    # via -r requirements.in
packaging==23.0
    # via mkdocs
proselint==0.13.0
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pygments==2.14.0
    # via mkdocs-material
pymdown-extensions==9.10
    # via
    #   -r requirements.in
    #   mkdocs-material
python-dateutil==2.8.2
    # via
    #   arrow
    #   ghp-import
pytz==2022.7.1
    # via mkdocs-git-revision-date-localized-plugin
pyyaml==6.0
    # via
    #   mkdocs
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
regex==2022.10.31
    # via mkdocs-material
requests==2.28.2
    # via mkdocs-material
six==1.16.0
    # via
    #   proselint
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sqlparse==0.4.3
    # via django
urllib3==1.26.15
    # via requests
uvicorn==0.21.1
    # via -r requirements.in
watchdog==2.3.1
    # via mkdocs
whitenoise==6.4.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
